move to catch instead of google test

+++ include/ocr/buffer.h
@@ -23,10 +23,10 @@ OCR_EXTERN_C_BEGIN
 
 typedef struct {
     u8 *data;
-    u64 size;
+    size_t size;
 } ocr_buffer_t;
 
-ocr_status_t ocr_buffer_init(ocr_buffer_t *buffer, ocr_pool_t *pool, u64 size);
+ocr_status_t ocr_buffer_init(ocr_buffer_t *buffer, ocr_pool_t *pool, size_t size);
 
 void ocr_buffer_destroy(ocr_buffer_t *buffer);
